Add files via upload
Updated version of the HTML code that generates a random background color every time the page is loaded for the first time.

In this updated code, the generateRandomColor function generates a random hexadecimal color value by randomly selecting characters from the set of alphanumeric characters (0-9, A-F). The setRandomBackgroundColor function is called during the initial page load to set the background color of the body element using the generated random color.